New Paths and Configuration Code
The Configuration now auto-generates if missing.
Sane Defaults are used, as well as comments.

The path code has been updated so that the ot_init.cfg
tell OT where:

The main config folder, the name of the config file.
(such as client.cfg or server.cfg)

The prefix folder.
The scripts folder.

The main config file now tells OT where the data folder,
as well as the name of the wallet.

This commit is made up of the following commits:

a549e8c NEW OT PATH: Reworked.
53da927 bugfix for ot prompt
fd88733 fixed build wrappers bug
7060e11 debuging...
0dc6879 bug fixes for linux
06ec449 updated project files for new version of java
112657c auto-generated config files now have comments
aab57e8 ft requested changes. + fix for knotwhat.
eb25183 more fixes for knotwhat
1339607 fix up global to member vars; also cleaned up some dll export
ifdefs
ef50f08 cleanup OpenTransactions.cpp
f60eac0 rebuilt wrappers
6ed45c8 moved conifg files back into config directory
4f454c7 half way there... rewriting the paths code
d76c004 ot prompt now working with new path code
95c53f1 finshed path re-write
15f0e3e re-build wrappers, api slightly changed, again
62de429 fixed a linker issues
5590fbd fixed compile on linux
7f9adb1 re_worked setting up header scripts
b5535e1 fixed issue when only having a foldername in OTDB

+++ src/otlib/OTAsymmetricKey.cpp
@@ -380,7 +380,7 @@ OPENSSL_CALLBACK_FUNC(default_pass_cb)
     const char * szFunc = "OPENSSL_CALLBACK_FUNC(default_pass_cb)";
     
 	int len = 0;
-    const uint32_t theSize = static_cast<const uint32_t>(size);
+    const uint32_t theSize = uint32_t(size);
 	// ------------------------------------
 	// We'd probably do something else if 'rwflag' is 1
 
@@ -415,7 +415,7 @@ OPENSSL_CALLBACK_FUNC(default_pass_cb)
 	const char *tmp_passwd = "test";
 //	const char *tmp_passwd = str_Password.c_str();
 
-	len = strlen(tmp_passwd);
+	len = static_cast<int> (strlen(tmp_passwd));
 //	len = str_Password.size();
 	
 	if (len <= 0)
@@ -704,7 +704,7 @@ OPENSSL_CALLBACK_FUNC(souped_up_pass_cb)
             // ---------------------------------------
             // The dialog should display this string (so the user knows what he is authorizing.)
             //
-            pCaller->SetDisplay(str_userdata.c_str(), str_userdata.size());
+            pCaller->SetDisplay(str_userdata.c_str(), static_cast<int> (str_userdata.size()));
             
             // ---------------------------------------
             if (1 == rwflag)
@@ -1972,7 +1972,7 @@ bool OTAsymmetricKey::LoadPublicKeyFromPGPKey(const OTASCIIArmor & strKey)
 	
 	BIO_get_mem_ptr(bio_out, &bptr);
 	
-	pgpKeys = ExportRsaKey((unsigned char*)bptr->data, bptr->length);
+	pgpKeys = ExportRsaKey((unsigned char*)bptr->data, static_cast<int> (bptr->length));
 	
 	if(!pgpKeys.pRsa)
 	{
